+++ SheepShaver/e2e/sse2e/__init__.py
@@ -1,0 +1,1 @@
+"""SheepShaver end-to-end VNC test harness (ROADMAP A5, P1)."""